backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
gunicorn==23.0.0
stripe==10.12.0
numpy==1.24.3
scikit-learn==1.3.0
geopy==2.3.0
Pillow==10.0.0
opencv-python==4.8.0.76

flask-login==0.6.3
flask-limiter==3.5.0
redis==5.0.1
celery==5.3.4
reportlab==4.0.7
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.6

flask-socketio==5.5.1

# Data analysis and business intelligence
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Additional dependencies for analytics and data pipeline
scipy==1.11.1
statsmodels==0.14.0

# Redis async support
async-timeout==4.0.3

# Security dependencies
pyotp==2.9.0
qrcode==7.4.2
email-validator==2.1.0
flask-jwt-extended==4.6.0
flask-talisman==1.1.0
flask-wtf==1.2.1

